refactor: change folders structure, rename components
