+++ src/pages/charactersPage/utils/getFormattedPagesArray.ts
