+++ src/pages/charactersPage/utils/getPagesArray.ts
